Simplify the prepull configuration
Rather than keeping the idea of multiple prepull configurations,
elevate the prepull configuration to the top level and assume
Rubin tag conventions.  We're not trying to be a generic prepuller,
we're specifically a controller for labs for the Notebook Aspect of
the Rubin Science Platform, so can take a much simpler approach.