Add proposal for lab creation for bot users
Add a mechanism for bot users to create labs via a POST to
JupyterHub without caring about what fields are present in the
lab creation form.

Also change the prepuller configuration to be a generic
configuration for what images are available to the Notebook
Aspect, which also happens to control prepulling.